Add Dockerfile for gcloud run deploy command
