+++ goblet/resources/pubsub.py
@@ -58,7 +58,7 @@ class PubSub(Handler):
         return
 
     def _deploy(self, sourceUrl=None, entrypoint=None, backend="cloudfunction"):
-        if not self.resources:
+        if not self.resources or not sourceUrl:
             return
 
         log.info("deploying topic functions......")
